wip: reorganize create workspace
