+++ src/kairos/workspace/adapter/repository/insert-workspace.ts
@@ -2,7 +2,7 @@ import { isDuplicateError } from '@/framework/mongodb/error/is-duplicate-error';
 import { WorkspacesCollection } from '@/framework/mongodb/schema/workspaces-collection-schema';
 import { err } from '@/kairos/shared/result/err';
 import { ok } from '@/kairos/shared/result/ok';
-import { type InsertWorkspace } from '@/kairos/workspace/application/insert-workspace.type';
+import { type InsertWorkspace } from '@/kairos/workspace/create-workspace/insert-workspace.type';
 import { workspaceSlugConflictError } from '@/kairos/workspace/errors';
 
 export const makeInsertWorkspace =
