+++ src/kairos/workspace/create-workspace/create-workspace.test.ts
@@ -1,7 +1,7 @@
 import { err } from '@/kairos/shared/result/err';
 import { ok } from '@/kairos/shared/result/ok';
-import { makeCreateWorkspace } from '@/kairos/workspace/application/create-workspace';
-import type { InsertWorkspace } from '@/kairos/workspace/application/insert-workspace.type';
+import { makeCreateWorkspace } from '@/kairos/workspace/create-workspace/create-workspace';
+import type { InsertWorkspace } from '@/kairos/workspace/create-workspace/insert-workspace.type';
 import { workspaceSlugConflictError } from '@/kairos/workspace/errors';
 import { describe, expect, test, vi } from 'vitest';
