+++ src/kairos/workspace/create-workspace/create-worskpace.compose.ts
@@ -1,7 +1,7 @@
 import { workspacesCollection } from '@/framework/mongodb/collection/workspaces-collection';
 import { makeInsertWorkspace } from '@/kairos/workspace/adapter/repository/insert-workspace';
-import { makeCreateWorkspace } from '@/kairos/workspace/application/create-workspace';
-import type { InsertWorkspace } from '@/kairos/workspace/application/insert-workspace.type';
+import { makeCreateWorkspace } from '@/kairos/workspace/create-workspace/create-workspace';
+import type { InsertWorkspace } from '@/kairos/workspace/create-workspace/insert-workspace.type';
 
 const insertWorkspace: InsertWorkspace = makeInsertWorkspace(workspacesCollection);
